Rework testcase modules for base functions
